Ensure project discovery is deterministic.
Each provider has a priority that we use to sort the enumeration. The provider with the highest priority wins. This ensures that if a "project.vs-chromium-project" file is dropped in a chromium enlistment, the file based project discovery provider wins. This allows scenarios where people want to disable file indexing in a Chromium enlistment.