Batch processing of file registration/unregistration events.
Registering (or Unregistering) a file takes a non trivial amount of CPU and Disk I/O time on the server process, due to cache invalidation, and potential to have to recompute the world. By batching those operations, and flushing them in a single operation, we spare lots of cycles on the server. This is useful, for example, when the user closes a lots of files from the Visual Studio editor (unregister), or when the user opens a lot of files (through a large refactoring or find/replace in files).