Implement search in "chunks" of 100KB
* In some cases, regex search performs badly overall, and hold a thread for a long time (i.e. a few seconds) when searching in big files. Holding a thread for that long makes the "Cancellation" of search operation less responsive, as we can only cancel when C# code has control.

For example, when searching for "[a-z]{40,100}" in a file about about 27MB in the Chromium code base ("third_party\cld_2\src\internal\cld2_generated_quad0122.cc"), regex search takes about 20sec. This essentially means that once started, it takes 20sec to cancel that particular search operation (not counting one of the thread CPU usage is at 100% during that time).

By splitting the search in chunks of 100KB, we aleviate the long cancellation problem. Of course, it would be nice to be able to distribute the work accross all available threads. It would be even nicer to have a regex implementation that is more efficient :).