Partial fix for issue #27
VsChromium incorrectly treats byte offsets as character offsets.
This works for simple ascii encoding, but creates issues with variable
length encoding, such as UTF8 and GBK.

To mitigate this problem, VsChromium now uses the (line, column)
information when navigating to a code search result. The
(line, column) pair is translated by the VS Editor into a character
offset.

This is not a 100% correct solution, as other part of VsChromium
also rely on the byte offset == character offset assumption, but
this CL fixes the most obvious user-facing issue (e.g. navigation).

Note that this issue does not apply to UTF16 files, as VsChromium
correctly translates character offset = byte offset / 2.