Support pausing while server is busy
This is so that we can stop the server if the server is
refreshing the index. We need the server to end up in a
manually requested pause state, so that the index is
100% refreshed on resume or manual refresh.

Also sprinkle more "CancellationToken" throughout the
file database refresh code, so that pausing is more
responsive.