Simplify thread pool to use BlockingCollection
Instead of implementing a custom queue with custom locks, CustomThreadPool
now uses a single BlockingCollection to add elements, and a fixed set
of threads consuming elements concurrently using GetConsumingEnumerable.

This fixes bug #60 and #61, at the expense of losing the ability of shutting
down idle threads in the pool. We decreased the pool size from 10 to 5.