Ensure in-order processing of certain IPC requests
The Server processes requests from VS, it uses a high priority
thread pool of 10 threads. This means there is a chance requests
may be processed out of order. This is fine for most requests,
and actually useful sometime (e.g. a new search request should
be able to run and cancel the previous one). However, there are
cases when requests should be processed in order. For example,
"RegisterFile"/"UnregisterFile" requests are used to keep track
of roots to detect project files. These requests need to be
processed in-order, otherwise the server might lose track of
some roots (or keep some roots active even they are not
open in VS anymore).

This is the 2nd part of fixing issue #64