Skip directories that cannot be enumerated
In some cases, for example when a junction point point to a file or
an invalid directory, GetDirectoryEntries throws a Win32 exception
"The directory name is invalid" (error code 267).

VsChromium is not more resilient to this type of errors, and simply
skip directories that cannot be enumerated for any reason, logging
a warning when it happens.

This fixes issue #66.