Ensures Resharper does not hang Visual Studio
When VsChromium indexes a large project, and "Solution Explorer" is enabled,
Resharper enumerates all entries (directories and files, recursively) shown
in Solution Explorer. For large projects, there can be hundred of thousand
directories, and even more files.

Resharper behavior leads to hangs in Visual Studio UI, because all
programmatic access to the Solution Explorer nodes happens on the UI
thread (IvsHierarchy).

This change implements a workaround in IVsHierarchy implementation
of Solution Explorer: If Resharper extension is detected in the list
of calling methods, we return E_NOTIMPL. This is clearly brittle, but
we could not figure out a better way.

This fixes issue #68