Use Running Document Table to track open documents
Instead of using the ITextDocumentFactoryService component, we now use
IVsRunningDocumentTable to keep track of when files are open and closed
in the VS editor.

The main reason is that ITextDocumentFactoryService keeps track of all
ITextBuffer instances used in VS, including ITextDocument instances
loaded for temporary usage, such as the "Find in File" features.

This change ensures the number of events related to opening and closing
documents is close to zero when using "Find in Files" on large directories
or project.

This is (hopefully) the last part to fix issue #61