Remove support for escape character in search pattern
Supporting '\' as the escape character creates confusion and
makes it difficult to support searching for strings containing
the '\' character. The only special character in search pattern
is not the wildcard ('*') character. There is no support for
escaping it. If this becomes a problem, there will be a need to
expose a UI option to switch between different escape modes, the
same way there are options to switch between regex or not.

This fixes issue #71